+++ src/cpp/core/include/core/Error.hpp
@@ -136,7 +136,7 @@ class ErrorLocation
 {
 public:
    ErrorLocation() ;
-   ErrorLocation(const char* function, const char* file, int line) ;
+   ErrorLocation(const char* function, const char* file, long line) ;
    virtual ~ErrorLocation() ;
    
    // immutable - copying and assignment via shared_ptr 
@@ -145,7 +145,7 @@ public:
 
    const std::string& function() const ;
    const std::string& file() const ;
-   int line() const ;
+   long line() const ;
    
    std::string asString() const;
